3_Encrption_Methods/test/pkg3_encrption_methods/AdditiveTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package pkg3_encrption_methods;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author lenova
 */
public class AdditiveTest {
    
    public AdditiveTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
    }
    
    @AfterClass
    public static void tearDownClass() {
    }
    
    @Before
    public void setUp() {
    }
    
    @After
    public void tearDown() {
    }

    /**
     * Test of encrypt method, of class Additive.
     */
    @Test
    public void testEncrypt() {
        System.out.println("encrypt");
        String strData = "";
        String expResult = "";
        String result = Additive.encrypt(strData);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of decrypt method, of class Additive.
     */
    @Test
    public void testDecrypt() {
        System.out.println("decrypt");
        String strData = "";
        String expResult = "";
        String result = Additive.decrypt(strData);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }
    
}
